pyproject.toml
[project]
name = "magic-conch-shell"
version = "0.1.0"
description = "Personal life brain + agent — answers to the shell."
requires-python = ">=3.13"

dependencies = [
  # CLI & UX
  "typer[all]>=0.24",
  "rich>=15",
  # File·데이터
  "watchdog>=6",
  "pydantic>=2.13",
  "pyyaml>=6",
  "python-frontmatter>=1.1",
  # HTTP·외부 API
  "httpx>=0.28",
  "notion-client>=3",
  # LLM
  "ollama>=0.6",           # 로컬 LLM 클라이언트
  "oauth-codex>=4",        # Codex OAuth (GPT-5.4)
  # MCP
  "fastmcp>=3",            # MCP 서버 프레임워크
  # Memory engine (Day 2 검증 후 최종 확정)
  "memsearch[ollama]>=0.3",
  # Time
  "tzdata>=2026.1",
]

[dependency-groups]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.25",
  "ruff>=0.8",
  "mypy>=1.15",
]

[project.scripts]
mcs = "mcs.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcs"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.mypy]
python_version = "3.13"
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
filterwarnings = [
  "ignore::DeprecationWarning:milvus_lite",
]
